Moving files into dexes/
